Refactor admin CLI: Add database, monitoring, and dev management commands
