+++ Makefile
@@ -155,6 +155,9 @@ admin-cli:
 	@echo "  python3 admin.py db migrate"
 	@echo "  python3 admin.py db status"
 	@echo "  python3 admin.py db test-connection"
+	@echo "  python3 admin.py db info"
+	@echo "  python3 admin.py db reset"
+	@echo "  python3 admin.py db clear-data"
 	@echo ""
 	@echo "User Management:"
 	@echo "  python3 admin.py user create-admin"
