+++ backend/admin.py
@@ -122,6 +122,160 @@ def db_downgrade(revision):
         sys.exit(1)
 
 
+def db_test_connection():
+    """Test database connection."""
+    try:
+        # Import database engine from cli.py approach
+        from sqlalchemy import create_engine, text
+        from backend.app.core.config import settings
+        
+        engine = create_engine(settings.database_url)
+        with engine.connect() as conn:
+            conn.execute(text("SELECT 1"))
+            print_success("Database connection successful")
+    except Exception as e:
+        print_error(f"Database connection failed: {e}")
+        sys.exit(1)
+
+
+def db_info():
+    """Show database information."""
+    try:
+        from sqlalchemy import create_engine, text
+        from backend.app.core.config import settings
+        
+        engine = create_engine(settings.database_url)
+        with engine.connect() as conn:
+            # Get database name
+            result = conn.execute(text("SELECT current_database()"))
+            db_name = result.scalar()
+            print_info(f"Database: {db_name}")
+
+            # Get table count
+            result = conn.execute(
+                text(
+                    """
+                SELECT COUNT(*) FROM information_schema.tables
+                WHERE table_schema = 'public'
+            """
+                ),
+            )
+            table_count = result.scalar()
+            print_info(f"Tables: {table_count}")
+
+            # Get user count
+            result = conn.execute(text("SELECT COUNT(*) FROM users"))
+            user_count = result.scalar()
+            print_info(f"Users: {user_count}")
+
+    except Exception as e:
+        print_error(f"Error getting database info: {e}")
+        sys.exit(1)
+
+
+def db_reset(confirm=False):
+    """Completely reset the database - drop all tables and recreate them."""
+    if not confirm:
+        response = input(
+            "⚠️  WARNING: This will completely delete all data in the database. "
+            "This action cannot be undone. Are you sure? (yes/no): ",
+        )
+        if response.lower() != "yes":
+            return
+
+    try:
+        from sqlalchemy import create_engine, text
+        from backend.app.core.config import settings
+        
+        engine = create_engine(settings.database_url)
+        
+        # Step 1: Drop all tables
+        with engine.connect() as conn:
+            # Disable foreign key checks temporarily
+            conn.execute(text("SET session_replication_role = replica;"))
+
+            # Get all table names
+            result = conn.execute(
+                text(
+                    """
+                SELECT tablename FROM pg_tables
+                WHERE schemaname = 'public'
+                AND tablename != 'alembic_version'
+            """
+                )
+            )
+            tables = [row[0] for row in result]
+
+            # Drop all tables
+            for table in tables:
+                conn.execute(text(f"DROP TABLE IF EXISTS {table} CASCADE"))
+
+            # Re-enable foreign key checks
+            conn.execute(text("SET session_replication_role = DEFAULT;"))
+            conn.commit()
+
+        # Step 2: Reset Alembic version table
+        with engine.connect() as conn:
+            conn.execute(text("DROP TABLE IF EXISTS alembic_version"))
+            conn.commit()
+
+        # Step 3: Run migrations to recreate all tables
+        db_migrate()
+
+        print_success("Database has been completely reset and reinitialized")
+
+    except Exception as e:
+        print_error(f"Database reset failed: {e}")
+        sys.exit(1)
+
+
+def db_clear_data(confirm=False):
+    """Clear all data from tables but keep the structure."""
+    if not confirm:
+        response = input(
+            "⚠️  WARNING: This will delete all data from all tables but keep the structure. "
+            "This action cannot be undone. Are you sure? (yes/no): ",
+        )
+        if response.lower() != "yes":
+            return
+
+    try:
+        from sqlalchemy import create_engine, text
+        from backend.app.core.config import settings
+        
+        engine = create_engine(settings.database_url)
+        
+        with engine.connect() as conn:
+            # Disable foreign key checks temporarily
+            conn.execute(text("SET session_replication_role = replica;"))
+
+            # Get all table names
+            result = conn.execute(
+                text(
+                    """
+                SELECT tablename FROM pg_tables
+                WHERE schemaname = 'public'
+                AND tablename != 'alembic_version'
+            """
+                )
+            )
+            tables = [row[0] for row in result]
+
+            # Truncate all tables
+            for table in tables:
+                conn.execute(text(f"TRUNCATE TABLE {table} RESTART IDENTITY CASCADE"))
+
+            # Re-enable foreign key checks
+            conn.execute(text("SET session_replication_role = DEFAULT;"))
+            conn.commit()
+
+        print_success("All data has been cleared from the database")
+
+    except Exception as e:
+        print_error(f"Data clearing failed: {e}")
+        sys.exit(1)
+
+
 def backup_create(output=None):
     """Create database backup."""
     if not output:
@@ -332,6 +486,36 @@ def monitoring_health():
         print_error(f"Database check error: {e}")
 
 
+def monitoring_logs(lines=50, level="INFO"):
+    """Show recent application logs."""
+    try:
+        from backend.app.core.config import settings
+        
+        log_file = settings.log_file
+        if not os.path.exists(log_file):
+            print_error(f"Log file not found: {log_file}")
+            return
+
+        with open(log_file) as f:
+            log_lines = f.readlines()
+
+        # Filter by level and get last N lines
+        filtered_lines = [line for line in log_lines if level in line]
+        recent_lines = (
+            filtered_lines[-lines:] if len(filtered_lines) > lines else filtered_lines
+        )
+
+        if recent_lines:
+            print_info(f"Recent {level} logs (last {len(recent_lines)} lines):")
+            for line in recent_lines:
+                print(line.rstrip())
+        else:
+            print_info(f"No {level} logs found in recent {lines} lines")
+
+    except Exception as e:
+        print_error(f"Error reading logs: {e}")
+
+
 def config_show():
     """Show current configuration."""
 
@@ -405,21 +589,58 @@ def dev_api_test(url="http://localhost:8000"):
         # Health check
         response = requests.get(f"{url}/health", timeout=5)
         if response.status_code == 200:
-            pass
+            print_success("Health check: OK")
         else:
-            pass
+            print_error(f"Health check failed: {response.status_code}")
 
         # API docs
         response = requests.get(f"{url}/docs", timeout=5)
         if response.status_code == 200:
-            pass
+            print_success("API docs: OK")
         else:
-            pass
+            print_error(f"API docs failed: {response.status_code}")
 
     except ImportError:
         print_error("API test failed: requests module not available")
     except Exception as e:
         print_error(f"API test failed: {e}")
+
+
+def dev_test_data(users=5):
+    """Create test data for development."""
+    try:
+        from sqlalchemy import create_engine
+        from sqlalchemy.orm import sessionmaker
+        from backend.app.core.config import settings
+        from backend.app.models.user import UserRole, UserStatus
+        from backend.app.schemas.user import UserCreate
+        from backend.app.services.user_service import UserService
+        
+        engine = create_engine(settings.database_url)
+        SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+        db = SessionLocal()
+        user_service = UserService(db)
+
+        for i in range(users):
+            user_data = UserCreate(
+                email=f"test{i}@example.com",
+                username=f"testuser{i}",
+                password="testpass123",
+                first_name=f"Test{i}",
+                last_name="User",
+                role=UserRole.USER,
+                status=UserStatus.ACTIVE,
+            )
+            user_service.create_user(user_data)
+
+        print_success(f"Created {users} test users successfully")
+
+    except Exception as e:
+        print_error(f"Error creating test data: {e}")
+        sys.exit(1)
+    finally:
+        if 'db' in locals():
+            db.close()
 
 
 def user_create_admin():
@@ -871,6 +1092,10 @@ def main():
     db_subparsers = db_parser.add_subparsers(dest="db_command")
     db_subparsers.add_parser("migrate", help="Run migrations")
     db_subparsers.add_parser("status", help="Show migration status")
+    db_subparsers.add_parser("test-connection", help="Test database connection")
+    db_subparsers.add_parser("info", help="Show database information")
+    db_subparsers.add_parser("reset", help="Completely reset database (drop all tables and recreate)")
+    db_subparsers.add_parser("clear-data", help="Clear all data but keep table structure")
 
     downgrade_parser = db_subparsers.add_parser(
         "downgrade",
@@ -959,6 +1184,10 @@ def main():
     monitoring_subparsers = monitoring_parser.add_subparsers(dest="monitoring_command")
     monitoring_subparsers.add_parser("health", help="Check system health")
     monitoring_subparsers.add_parser("containers", help="Monitor containers")
+    
+    logs_parser = monitoring_subparsers.add_parser("logs", help="Show logs")
+    logs_parser.add_argument("--lines", type=int, default=50, help="Number of lines")
+    logs_parser.add_argument("--level", default="INFO", help="Log level")
 
     # Config commands
     config_parser = subparsers.add_parser("config", help="Configuration management")
@@ -969,6 +1198,14 @@ def main():
     # Dev commands
     dev_parser = subparsers.add_parser("dev", help="Development tools")
     dev_subparsers = dev_parser.add_subparsers(dest="dev_command")
+
+    test_data_parser = dev_subparsers.add_parser("test-data", help="Create test data")
+    test_data_parser.add_argument(
+        "--users",
+        type=int,
+        default=5,
+        help="Number of users",
+    )
 
     dev_subparsers.add_parser("quality-check", help="Run code quality checks")
 
@@ -1001,6 +1238,14 @@ def main():
             db_status()
         elif args.db_command == "downgrade":
             db_downgrade(args.revision)
+        elif args.db_command == "test-connection":
+            db_test_connection()
+        elif args.db_command == "info":
+            db_info()
+        elif args.db_command == "reset":
+            db_reset()
+        elif args.db_command == "clear-data":
+            db_clear_data()
 
     elif args.command == "user":
         if args.user_command == "create-admin":
@@ -1052,6 +1297,8 @@ def main():
             monitoring_health()
         elif args.monitoring_command == "containers":
             monitoring_containers()
+        elif args.monitoring_command == "logs":
+            monitoring_logs(args.lines, args.level)
 
     elif args.command == "config":
         if args.config_command == "show":
@@ -1060,7 +1307,9 @@ def main():
             config_validate()
 
     elif args.command == "dev":
-        if args.dev_command == "quality-check":
+        if args.dev_command == "test-data":
+            dev_test_data(args.users)
+        elif args.dev_command == "quality-check":
             dev_quality_check()
         elif args.dev_command == "api-test":
             dev_api_test(args.url)
